Se edito la el archivo persona provaider para consumir el servicio de azure desde la app

+++ lib/widgets/alertas.dart
@@ -3,7 +3,7 @@ import 'package:tts/tts.dart';
 
 class Alertas {
   void showDialogMessage(BuildContext context, IconData icono, Color color,
-      String title, int edad, subtitle) {
+      String title, double edad, subtitle) {
     showDialog(
       context: context,
       builder: (BuildContext context) {
